feat: card component displays correct info test
